refactor: move buildBalanceOperations to transaction handler helper :hammer:
