+++ components/ledger/internal/adapters/http/in/transaction_helper.go
@@ -5,9 +5,14 @@
 package in
 
 import (
+	"sort"
 	"strings"
 
+	"github.com/LerianStudio/midaz/v3/pkg/constant"
+	"github.com/LerianStudio/midaz/v3/pkg/mmodel"
 	"github.com/LerianStudio/midaz/v3/pkg/net/http"
+	pkgTransaction "github.com/LerianStudio/midaz/v3/pkg/transaction"
+	"github.com/LerianStudio/midaz/v3/pkg/utils"
 
 	"github.com/gofiber/fiber/v2"
 	"github.com/google/uuid"
@@ -72,4 +77,67 @@ func getAliasWithoutKey(array []string) []string {
 	}
 
 	return result
+}
+
+// buildBalanceOperations constructs and sorts balance operations from the
+// validated transaction entries. This is pure logic with no I/O dependencies.
+// Operations are sorted by internal key to prevent deadlocks in the Lua script.
+func buildBalanceOperations(organizationID, ledgerID uuid.UUID, validate *pkgTransaction.Responses, balances []*mmodel.Balance) []mmodel.BalanceOperation {
+	balanceOperations := make([]mmodel.BalanceOperation, 0)
+
+	for _, balance := range balances {
+		balanceKey := balance.Key
+		if balanceKey == "" {
+			balanceKey = constant.DefaultBalanceKey
+		}
+
+		aliasKey := balance.Alias + "#" + balanceKey
+		internalKey := utils.BalanceInternalKey(organizationID, ledgerID, aliasKey)
+
+		for k, v := range validate.From {
+			if pkgTransaction.SplitAliasWithKey(k) == aliasKey {
+				if pkgTransaction.IsDoubleEntrySource(v) {
+					op1, op2 := pkgTransaction.SplitDoubleEntryOps(v)
+
+					balanceOperations = append(balanceOperations, mmodel.BalanceOperation{
+						Balance:     balance,
+						Alias:       k,
+						Amount:      op1,
+						InternalKey: internalKey,
+					})
+
+					balanceOperations = append(balanceOperations, mmodel.BalanceOperation{
+						Balance:     balance,
+						Alias:       k,
+						Amount:      op2,
+						InternalKey: internalKey,
+					})
+				} else {
+					balanceOperations = append(balanceOperations, mmodel.BalanceOperation{
+						Balance:     balance,
+						Alias:       k,
+						Amount:      v,
+						InternalKey: internalKey,
+					})
+				}
+			}
+		}
+
+		for k, v := range validate.To {
+			if pkgTransaction.SplitAliasWithKey(k) == aliasKey {
+				balanceOperations = append(balanceOperations, mmodel.BalanceOperation{
+					Balance:     balance,
+					Alias:       k,
+					Amount:      v,
+					InternalKey: internalKey,
+				})
+			}
+		}
+	}
+
+	sort.Slice(balanceOperations, func(i, j int) bool {
+		return balanceOperations[i].InternalKey < balanceOperations[j].InternalKey
+	})
+
+	return balanceOperations
 }
